update controller to use config and chat session
