+++ packages/server/api/src/app/ai/ai.module.ts
@@ -1,6 +1,6 @@
 import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
-import { aiChatStatelessController } from './chat/ai-chat-stateless.controller';
 import { aiChatController } from './chat/ai-chat.controller';
+import { aiToolChatController } from './chat/ai-tool-chat.controller';
 import { aiConfigController } from './config/ai-config.controller';
 import { mcpToolsController } from './mcp-client/mcp-tools.controller';
 import { aiProvidersController } from './providers/ai-providers.controller';
@@ -22,7 +22,7 @@ export const aiModule: FastifyPluginAsyncTypebox = async (app) => {
     prefix: '/v1/ai/mcp-tools',
   });
 
-  await app.register(aiChatStatelessController, {
+  await app.register(aiToolChatController, {
     prefix: '/v1/ai/chat-stateless',
   });
 };
